fix: isolate OpenAPI egress DNS checks
